feaT: use hashring for ratleimits

+++ apps/agent/pkg/ratelimit/interface.go
@@ -6,15 +6,15 @@ type Ratelimiter interface {
 
 type RatelimitRequest struct {
 	Identifier     string
-	Max            int32
-	Cost           int32
-	RefillRate     int32
-	RefillInterval int32
+	Max            int64
+	Cost           int64
+	RefillRate     int64
+	RefillInterval int64
 }
 
 type RatelimitResponse struct {
 	Pass      bool
-	Limit     int32
-	Remaining int32
+	Limit     int64
+	Remaining int64
 	Reset     int64
 }
